+++ NEWS.md
@@ -4,17 +4,27 @@
 * Potential evaporation coefficients can now be estimated from soil properties.
 * Rooting profiles can now be estimated with an equation and parameters
   for each vegetation type.
+* New option to select method for surface albedo:
+    * fixed, cover-weighted sum over PFTs and bare ground with fixed values
+    * dynamic (new), albedo components for vegetation (LAI dependent),
+      soil (moisture darkening) and snow (age- and temperature dependent)
 
 ## Changes to interface
-* The class `"swSite"` gained slot `"PotSoilEvCoMethod"` and
-  associated method `swSite_PotSoilEvCoMethod()`.
-  It also gained slot `"RootingProfileMethod"` and
-  associated methods `swSite_RootingProfileMethod()`.
-* The class `"swProd2"` gained slot `"RootProfileParameters"`.
+* The class `"swSite"` gained
+    * slot `"PotSoilEvCoMethod"` and methods `swSite_PotSoilEvCoMethod()`
+    * slot `"RootingProfileMethod"` and methods `swSite_RootingProfileMethod()`
+    * slot `"AlbedoMethod"` and methods `swSite_AlbedoMethod()`
+    * additional new slots `"RoughnessLengthGroundSurface"`, `"AlbedoSnowMax"`,
+      `"AlbedoSoilDry"`, `"AlbedoSoilSaturated"`,
+      `"AlbedoSoilDarkeningParameter"`, `"SnowFractionalCoverMeltingFactor"`
+* The class `"swProd2"` gained slots `"RootProfileParameters"` and
+  `"kExtVegAlbedo"`.
 
 * The class `"swOutput"` produces now output of soil evaporation for
   each soil layer (previously, output was produced only for soil layers with
   soil evaporation potential).
+* The class `"swOutput"` now contains a new output element `"SW_ENERGYAVG"`
+  with output of surface albedo.
 
 
 # rSOILWAT2 v6.5.2